test(cli): lighten Windows service recovery coverage
